* dev: (155 commits)
  DEFEDIT-614 Minor fixes after review
  DEF-2357 merge branch (#1161)
  added releasenotification@king.com
  Moved bezier calculations outside loop in RigUtil.java. (#1159)
  DEFEDIT-614 Typo in test comment
  DEFEDIT-614 Fixed issue where new "unknown" files can't be opened
  DEFEDIT-614 Tabs closes when the underlying resources are removed
  DEFEDIT-614 Removed undo test for node deletion and views, not relevant anymore
  DEFEDIT-614 Removed resource node ids from tile map and tile source editors
  DEF-2370 Fixed arch names for iOS in build.yml (#1156)
  DEFEDIT-614 Fixed issues with properties view being blank
  Native Extension: Added support for osx (#1130)
  Missing property documentation added.
  DEFEDIT-654 Review fixes
  DEFEDIT-614 Fixed outline, which had broken after tabs did not retain resource nodes
  DEFEDIT-654 Fix build error navigation
  Bumped version to 1.2.96
  DEF-2348 - Exposed missing spine gui node controls. (#1140)
  DEF-2343 - Fixed issue with spine scenes that had both rotation and scaling in bind pose. (#1139)
  DEF-2358 Fixed issue with using go.set with vector4 on Label component (#1147)
  ...

# Conflicts:
#	editor/styling/stylesheets/components/_text-area.scss
#	editor/styling/stylesheets/editor.scss
#	editor/styling/stylesheets/main.sass

+++ com.dynamo.cr/com.dynamo.cr.bob/src/com/dynamo/bob/Project.java
@@ -37,6 +37,7 @@ import com.dynamo.bob.bundle.IOSBundler;
 import com.dynamo.bob.bundle.LinuxBundler;
 import com.dynamo.bob.bundle.OSXBundler;
 import com.dynamo.bob.bundle.Win32Bundler;
+import com.dynamo.bob.bundle.Win64Bundler;
 import com.dynamo.bob.fs.ClassLoaderMountPoint;
 import com.dynamo.bob.fs.FileSystemWalker;
 import com.dynamo.bob.fs.IFileSystem;
@@ -365,6 +366,7 @@ public class Project {
         bundlers.put(Platform.X86Linux, LinuxBundler.class);
         bundlers.put(Platform.X86_64Linux, LinuxBundler.class);
         bundlers.put(Platform.X86Win32, Win32Bundler.class);
+        bundlers.put(Platform.X86_64Win32, Win64Bundler.class);
         bundlers.put(Platform.Armv7Android, AndroidBundler.class);
         bundlers.put(Platform.Armv7Darwin, IOSBundler.class);
         bundlers.put(Platform.JsWeb, HTML5Bundler.class);
@@ -743,14 +745,32 @@ run:
                 if (code == 304) {
                     // Reusing cached library
                 } else {
-                    input = new BufferedInputStream(connection.getInputStream());
-                    FileUtils.copyInputStreamToFile(input, f);
-                    try {
-                        ZipFile zip = new ZipFile(f);
-                        zip.close();
-                    } catch (ZipException e) {
-                        f.delete();
-                        throw new LibraryException(String.format("The file obtained from %s is not a valid zip file", url.toString()), e);
+                    boolean serverSha1Match = false;
+                    if(code == 200) {
+                        // GitHub uses eTags and we can check we have the up to date version by comparing SHA1 and server eTag if we get a 200 OK response
+                        if(sha1 != null)
+                        {
+                            String serverETag = connection.getHeaderField("ETag");
+                            if(serverETag != null)
+                            {
+                                if(sha1.equals(serverETag.replace("\"", ""))) {
+                                    // Reusing cached library
+                                   serverSha1Match = true;
+                                }
+                            }
+                        }
+                    }
+                    if(!serverSha1Match) {
+                        input = new BufferedInputStream(connection.getInputStream());
+                        FileUtils.copyInputStreamToFile(input, f);
+
+                        try {
+                            ZipFile zip = new ZipFile(f);
+                            zip.close();
+                        } catch (ZipException e) {
+                            f.delete();
+                            throw new LibraryException(String.format("The file obtained from %s is not a valid zip file", url.toString()), e);
+                        }
                     }
                 }
                 connection.disconnect();
@@ -759,7 +779,9 @@ run:
             } catch (FileNotFoundException e) {
                 throw new LibraryException(String.format("The URL %s points to a resource which doesn't exist", url.toString()), e);
             } finally {
-                IOUtils.closeQuietly(input);
+                if(input != null) {
+                    IOUtils.closeQuietly(input);
+                }
                 subProgress.worked(1);
             }
         }
